+++ libcovebods/common_checks.py
@@ -1,5 +1,5 @@
 from libcove.lib.common import common_checks_context
-from libcovebods.lib.common_checks import get_statistics
+from libcovebods.lib.common_checks import get_statistics, RunAdditionalChecks
 
 
 def common_checks_bods(context, upload_dir, json_data, schema_obj):
@@ -8,8 +8,12 @@ def common_checks_bods(context, upload_dir, json_data, schema_obj):
 
     context.update(common_checks['context'])
 
+    additional_checks = RunAdditionalChecks(json_data).run()
+
     context.update({
-      'statistics': get_statistics(json_data)
+        'statistics': get_statistics(json_data),
+        'additional_checks': additional_checks,
+        'additional_checks_count': len(additional_checks),
     })
 
     return context