+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Additional Checks: All relevant statements are present
+
 ## [0.1.0] - 2018-11-22
 
 First Release